workflow-platform/backend/requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.30,<1.0
SQLAlchemy>=2.0,<3.0
pydantic>=2.8,<3.0
PyYAML>=6.0,<7.0
python-multipart>=0.0.9,<1.0
pytest>=8.2,<9.0
httpx>=0.27,<1.0
